+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Changed "Fling Upwards" effect to force the player to stop grappling.
+
 ## [2.0.6] - 2023-09-29
 
 - Changed "Reforge Item" effect to be able to reforge any valid item.