+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Changed "Fling Upwards" effect to force the player to stop grappling.
+- Changed Challenge effects to give 5 Gold Coins when completed successfully.
 
 ## [2.0.6] - 2023-09-29