+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed "Noclip" effect to prevent item use whilst being shimmered.
 - Changed "Noclip" effect to continue shimmering whilst in lava.
 - Changed "Random Teleport" effect to ignore the Demon Conch if the world's spawn is in Hell.
+- Changed "Catch Critter" challenge to spawn 2 critters nearby, closer than before.
 
 ## [2.0.6] - 2023-09-29